chore: move get distance to utils
