+++ src/components/app/home/Cards/Cards.tsx
@@ -1,29 +1,13 @@
 import { useEffect, useState } from 'react';
 import { TbLocationFilled } from 'react-icons/tb';
 import { IoMdTrendingUp } from 'react-icons/io';
-import { cn } from '@/libs/utils';
+import { cn, getDistance } from '@/libs/utils';
 import type { Location } from '../../map/types';
 import PandalCard from './PandalCard';
 
 interface CardsProps {
   pandals: Location[];
 }
-
-const getDistance = (lat1: number, lng1: number, lat2: number, lng2: number) => {
-  const toRad = (value: number) => (value * Math.PI) / 180;
-  const R = 6371; // Earth's radius in kilometers
-
-  const dLat = toRad(lat2 - lat1);
-  const dLng = toRad(lng2 - lng1);
-
-  const a =
-    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-    Math.cos(toRad(lat1)) * Math.cos(toRad(lat2)) * Math.sin(dLng / 2) * Math.sin(dLng / 2);
-
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-
-  return R * c; // Distance in kilometers
-};
 
 const Cards = ({ pandals }: CardsProps) => {
   const [activeCard, setActiveCard] = useState<'trending' | 'nearme'>('trending');
